+++ Changes.md
@@ -2,6 +2,8 @@ Version Information
 ===================
 Version 3.3.0.1
   1. Update to Moodle 3.3.
+  2. Update to core chart API.
+  3. Add continue button, on / off setting (admin) and colour settings for admin and course.
 
 Version 3.1.1.1
   1. Update to latest ChartistJS 0.10.1 and Fill Donut plugin: